app/requirements.txt
Flask>=3.0
flasgger>=0.9.7
gunicorn>=21.2

opentelemetry-sdk>=1.24
opentelemetry-exporter-otlp-proto-http>=1.24
opentelemetry-instrumentation-flask>=0.47b0
opentelemetry-instrumentation-requests>=0.47b0
